+++ server/__init__.py
@@ -1,0 +1,27 @@
+import os
+
+from flask import Flask
+
+from server.logger import logger
+
+ROOT_PATH = os.path.dirname(os.path.realpath(__file__))
+
+LOG = logger.get_root_logger(os.environ.get(
+    'ROOT_LOGGER', 'root'), filename=os.path.join(ROOT_PATH, 'output.log'))
+
+
+def create_app():
+    LOG.info('running environment: %s', os.environ.get('ENV'))
+    # Debug mode if development env
+    app = Flask(__name__, instance_relative_config=True)
+    app.config.from_pyfile('config.py')
+
+    from .views import bp
+    app.register_blueprint(bp)
+
+    return app
+
+
+if __name__ == '__main__':
+    app = create_app()
+    app.run(host='0.0.0.0', port=app.config.get('PORT', 5000))  # Run the app
