--- server/template/index.html
@@ -1,10 +1,0 @@
-<!doctype html>
-<head>
-    <title>Testing HTML2</title>
-    <link rel=stylesheet type=text/css href="{{ url_for('static', filename='style.css') }}">
-</head>
-<body>
-<a class="user" href="{{ url_for('user') }}">
-    list users
-</a>
-</body>
